GDPR_RAG/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
entrypoints==0.4
executing==2.0.1
Faker==24.4.0
favicon==0.7.0
filelock==3.13.1
fonttools==4.49.0
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
html2text==2020.1.16
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanize==4.9.0
hyperframe==6.0.1
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
jax==0.4.25
jaxlib==0.4.25
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.38
langchain-text-splitters==0.0.1
langsmith==0.1.38
lazy_loader==0.3
llama-hub==0.0.79.post1
llama-index==0.10.26
llama-index-agent-openai==0.2.1
llama-index-cli==0.1.11
llama-index-core==0.10.26
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-github==0.1.5
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-qdrant==0.1.1
llama-parse==0.4.0
llamaindex-py-client==0.1.15
lxml==5.2.1
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
minijinja==1.0.16
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.16.1
opt-einsum==3.3.0
orjson==3.10.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.12.0
pyarrow==15.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7.1
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.1
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.7.3
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rich==13.7.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit==1.31.1
streamlit-aggrid==1.0.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.39.3
trulens==0.13.4
trulens-eval==0.27.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
utils==1.0.2
validators==0.22.0
wcwidth==0.2.13
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
